scripts/codelist/upload/requirements.txt
boto3==1.26.41
python-dotenv